Scientific_papers_summariser(Alaa_hussien)/requirements.txt
crewai
pydantic
serpapi
python-dotenv
beautifulsoup4
requests
